BUG FIX (partial): Addressed UI lag of engagement interactions/icons between user auth switch
Previously, when the user logged-out of their account, then navigated to a page with engagement icons (e.g. thumbs-up/down, favourite button, etc.), their actions still appeared unless the window was refreshed. For example, if they "liked" a comment, the thumbs-up icon is shown as "filled", but if they logout and navigate back to the page (without refreshing the window), the filled icon is still displayed despite their now un-authorized access. These changes addressed this problem for ALL engagement icons/actions except for the ones on the Blog Post page.